bookmarks: granular notebook file operations
This commit updates the Bookmarks component such that creating and
deleting bookmarks do not require a save of the buffer.

Prior to this, a single overwriting write operation would sync the tree
of bookmarks to disk.

Now, appends and selective removal of bookmarks from the bookmark files
is possible, allowing for the creation and deletion of bookmarks without
writing all dirty bookmarks to disk.

Signed-off-by: Louis DeLosSantos <louis.delos@gmail.com>