useability,commits: open local buffer when commit is head
prior to this commit opening a diff in the Commits component would
always open an in-memory text blob from the `git-show`.

now, if Commits is asked to open a diff for a file, and the commit is
the same as HEAD, the actual file is opened on the right side of
the diff, allowing for you to close side a, and immediately edit the
diff'd file.

Signed-off-by: Louis DeLosSantos <louis.delos@gmail.com>