useability: handle last win close issues
Last win close issues are annoying scenarios when you close the last
source code buffer.

This happens during diffs and when closing the last window accidentally.

For diffs, we now split a window when the last buffer to use is no
longer valid.

For source code buffers, we refuse to close the very last source code
window if nvim-ide component windows are visible. You can hide nvim-ide
component windows to close with ':q' as normal.

Forcefully closing with ':qa!' will work as normal

Signed-off-by: ldelossa <louis.delos@gmail.com>